C68k/SCSP fixes 1/6: Optimize the M68k critical path
git-svn-id: https://yabause.svn.sourceforge.net/svnroot/yabause/trunk@1970 d7be59e3-070e-4b9a-85a1-4646befae34d

+++ yabause/src/scsp.h
@@ -60,7 +60,7 @@ typedef struct
 typedef struct
 {
   u32 scsptiming1;
-  float scsptiming2;
+  u32 scsptiming2;  // 16.16 fixed point
 
   m68kcodebreakpoint_struct codebreakpoint[MAX_BREAKPOINTS];
   int numcodebreakpoints;
